Added first upgrade and ability to go back to the main menu
Pressing the back button brings you back to the main menu.
The first upgrade is shooting in three directions (front and both
sides). This is easily controlled by a boolean value which could be set
by device data.
There is an error right now which the side bolts do not shoot as far
when the ship is moving side to side. (not sure where the problem lays
exactly but we should come back to that)